// src/api/model/getMapsParams.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * MaCarte API
 * Documentation OpenAPI de l'API MaCarte
 * OpenAPI spec version: 1.0.0
 */
import type { OffsetParameter } from './offsetParameter';

export type GetMapsParams = {
/**
 * Contexte de la recherche. Doit être parmi :
 * - 'atlas' : recherche publique parmi les cartes partagées
 * - 'organization' : recherche des cartes publiques & atlas de l'organisation
 * - 'profile' : recherche parmi les cartes de l'utilisateur (connexion requise)
 *
 * Une carte partagée a share='atlas', valid=true et active=true.
 *
 * **Ajouter le header 'Authorization: Bearer xxx' pour le cas 'profile'.** pour ne pas avoir les erreurs 401 et 403.
 */
context?: string;
/**
 * Un ou plusieurs mots que doivent contenir le titre, la description ou le thème. S'il y a plusieurs mots, la recherche les interprète comme un ET.
 */
query?: string;
/**
 * Thème des cartes <br>
 *                     - nom du thème  <br>
 *                     - 'undefined' pour avoir les cartes qui n'ont pas de thème <br>
 *                     - null pour avoir toutes les cartes, indépendemment du thème
 */
theme?: string;
/**
 * Type des cartes ('macarte', 'storymap')
 */
type?: string;
/**
 * Premium des cartes ('default', 'edugeo')
 */
premium?: string;
/**
 * Identifiant public de l'organisation
 */
organization?: string;
/**
 * Nom public du créateur des cartes
 */
user?: string;
/**
 * Si context 'profile' : filtre les cartes actives (true) ou non-actives (false). Si context 'atlas', active=true.
 */
active?: boolean;
/**
 * Si context 'profile' : filtre les cartes valides (true) ou invalides (false). Si context 'atlas', valid=true.
 */
valid?: boolean;
/**
 * Si context 'profile' : filtre les cartes publiques ('atlas') ou privées ('private'). Si context 'atlas', share='atlas'.
 */
share?: string;
/**
 * Tri des cartes ('date', 'rank', 'views')
 */
sort?: string;
/**
 * Nombre maximum de cartes à retourner. Peut être 'all' dans le context 'profile'. Si limit > 100, limit sera considérée comme 100.
 */
limit?: number | string;
/**
 * Le nombre d'objets à ignorer avant de commencer à collecter l'ensemble de résultats
 */
offset?: OffsetParameter;
};